ORS WAR instance

This R6 class is currently non-functional. I am currently not aware of any reliable way to assert that Tomcat and JRE are up and running. If you can help, please consider opening a pull request here: https://github.com/jslth/rors.

Creates a new ORS instance object that operates on a WAR instance of OpenRouteservice, i.e. an instance that is started using a .war file. This R6 class is typically constructed by ors_instance. ORSJar requires Tomcat >= 10 to be installed on the system. Tomcat requires a Java Runtime Environment (JRE) to run. Additionally, both JAVA_HOME and CATALINA_HOME environment variables must be set to start Tomcat. If these requirements cannot be met, consider using ORSDocker or ORSJar.

For technical details on the setup of local ORS instances, refer to the Running WAR documentation. For details on how to use ORSDocker objects, refer to the installation vignette:

vignette("ors-installation", package = "rors")

Super classes

rors::ORSInstance -> rors::ORSLocal -> ORSWar

Methods

Public methods

ORSWar$new()

ORSWar$clone()

Inherited methods

rors::ORSInstance$get_status()

rors::ORSInstance$get_url()

rors::ORSInstance$is_mounted()

rors::ORSInstance$is_ready()

rors::ORSLocal$add_profiles()

rors::ORSLocal$report()

rors::ORSLocal$rm_extract()

rors::ORSLocal$rm_profiles()

rors::ORSLocal$set_endpoints()

rors::ORSLocal$set_extract()

rors::ORSLocal$update()

Method new()

Initialize the ORSWar object.

Usage

ORSWar$new(dir, version = NULL, overwrite = FALSE, verbose = TRUE, ...)

Arguments

dir

[character]

Custom OpenRouteService directory. If not specified, the war file will be downloaded to the current working directory. If a directory called "openrouteservice-{version}" is present, the download will be skipped. Ignored if server is not NULL.

version

[character]

The OpenRouteService version to use. Can either be a version number (e.g. 8.1.1) or "master". Defaults to the most recent supported version.

overwrite

[logical]

Whether to overwrite the current OpenRouteService directory if it exists.

verbose

[logical]

Level of verbosity. If TRUE, shows informative warnings and messages, spinners, progress bars and system notifications.

...

Not used.

Method clone()

The objects of this class are cloneable with this method.

Usage

ORSWar$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.