ORS-based sampling

Sample point geometries within the boundaries of the running ORS extract.

If the mounted instance is not local, the running ORS extract cannot be identified. In this case, a random bounding box is generated. See get_extract_boundaries for details.

Usage

ors_sample(
  size,
  ...,
  force_new_extract = FALSE,
  instance = NULL,
  poly = NULL,
  verbose = TRUE
)

Arguments

size

[integer]

Number of points to be sampled.

...

Passed to st_sample.

force_new_extract

[logical]

If TRUE, forces the cached extract geometries to be overwritten. Defaults to FALSE to increase speed. Set this to TRUE if the extract has changed and a new one needs to be loaded.

instance

[ors_instance]

Object of an OpenRouteService instance that should be used for route computations. It is recommended to use ors_instance to set an instance globally. This argument should only be used if activating an instance globally is not feasible.

poly

[sf/sfc]

Boundary polygon used to sample points. If NULL, the default, boundary polygons are extracted from the current instance as set by ors_instance.

verbose

[logical]

If TRUE, prints a loading spinner.

Value

sfc object containing the sampled POINT geometries

Details

The function vectortranslates the extract file to an sf object. Depending on the file, this can take a while. The unionized boundaries are then cached making subsequent function calls a lot faster.