Parameter info

Get information on the checks and preparations performed by ors_params for additional ORS parameters.

Usage

param_info()

Value

A dataframe or tibble with the following columns:

name

Name of the parameter

parent

Parent object of the parameter if it is nested, NA otherwise

match

Does the parameter have a predefined list of values?

scalar

Should the parameter be length-1?

type

R type of the parameter value

lim

Is parameter subject to numerical limits?

allowed

Which profiles are required to use the parameter?

poly

Should the parameter be formatted as a polygon?

box

Does the parameter need boxing before POSTing?

inspect

Is the parameter available for ors_inspect?

pariwise

Is the parameter available for ors_pairwise?

accessibility

Is the parameter available for ors_accessibility?