rors: Routing using 'OpenRouteService'

Flexible and opinionated interface to local and remote 'OpenRouteService' instances. The package enables qualitative and quantitative route computations, accessbility analyses, distance matrix calculations, street snapping and graph network exports. It is primarily aimed at processing larger datasets with controlled server-side limitations. R6 classes support the setup and and management of custom 'OpenRouteService' instances. The package functions take a tidy and graceful approach in querying and formatting the API output.

Usage

To learn more about the usage of this package, look into the two vignettes for this package:

vignette("ors-installation", package = "rors")
vignette("ors-routing", package = "rors")

Package options

The package relies on a number of package-level options that can be customized:

rors_setup_warn

If TRUE, emits warnings when a log entry of type "WARN" is encountered during the ORS setup. This concerns ORS setups using the ORSLocal class. Defaults to TRUE.

rors_echo

If TRUE, prints the results of a dry run (using req_dry_run) to the console each time a request is sent to the mounted ORS server. Useful for debugging or understanding what happens behind the scenes.

rors_throttle_directions, rors_throttle_isochrones, rors_throttle_matrix, rors_throttle_snap

Throttle specifications for each ORS service. throttle indicates the number of allowed requests per minute that can be sent to the server. The defaults can be retrieved from the ORS plans. Throttle values can either be a decimal number of (preferably) a fraction, e.g. 40 / 60. These options are useful for exploiting rate limits of the collaborative plan or other ORS servers that apply rate limits.

rors_retries

Specifies the number of retries that are performed before signaling an HTTP error. This option can be useful if the connection to a server is unstable. Defaults to 3 retries.

References

This package is powered by OpenRouteService. For problems concerning their service, refer to https://github.com/GIScience/openrouteservice

© openrouteservice.org by HeiGIT | Map data © OpenStreetMap contributors

See also

Author

Maintainer: Jonas Lieth jonas.lieth@gesis.org [copyright holder]