ORS instance

This R6 class is the foundation for R6 classes ORSLocal and ORSRemote and cannot be initialized. It defines methods to check status and information of a remote ORS server and to mount itself to the R session.

Methods

Public methods

ORSInstance$is_ready()

ORSInstance$get_status()

ORSInstance$is_mounted()

ORSInstance$clone()

Method is_ready()

Checks if ORS is ready to use. If TRUE, the routing functions can be used.

Usage

ORSInstance$is_ready()

Method get_status()

Retrieves information about the ORS server.

Usage

ORSInstance$get_status()

Method is_mounted()

Checks whether the instance is currently mounted to the session. More technically, $is_mounted compares the MD5 hashes of its R6 object and the currently mounted instance. Hence, an instance only counts as mounted if it is exactly the same as the currently mounted instance. An instance that was manually manipulated and has not been updated with $update() thus is not considered mounted.

Usage

ORSInstance$is_mounted()

Method clone()

The objects of this class are cloneable with this method.

Usage

ORSInstance$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.