ORS local instance

Creates a new ORS instance object. This R6 class is typically constructed by ors_instance.

For technical details on the setup of local ORS instances, refer to the official ORS reference. For details on how to use ORSLocal objects, refer to the installation vignette:

vignette("ors-installation", package = "rors")

Details

ORSLocal defines methods for all four steps in the ORS setup:

Extract: Download an extract and set it up for graph building.

Compose: Change docker settings that control how the service is set up.

Configuration: Change the way ORS computes routes.

Docker: Send commands to docker to control the ORS container.

Docker commands

Local ORS instances are built using Docker containers. Initializing ORSLocal downloads a docker-compose.yml file that tells Docker how to build an ORS container. The Docker backend is pulled as a Docker image (see Dockerhub). rors communicates with Docker from within R to setup and manage ORS containers.

On Unix systems, Docker requires superuser permissions. In Unix shells, this is not problematic. However, R cannot communicate with Docker without explicitly being granted superuser permissions or at least permissions to access Docker. Thus, ORSLocal checks if the current user is the superuser or if the current user has access to the Docker group. If not, it aborts. For more details on how to manage Docker as a non-root user on Linux, refer to the Docker documentation. Note that this procedure grants root-level privileges to a user and can negatively impact your system's security.

Jumpstarting

When first initializing ORSLocal or when first creating a new directory, ORSLocal will ask you if you would like to perform a jumpstart. Jumpstarting means creating an initial Docker build that runs on default settings, config and an extract file from Heidelberg, Germany. On the initial build, Docker also creates the file system of ORS instances including a configuration file. Irrespective of your choice to jumpstart, Docker only allows one routing profile and automatically uses the default extract of Heidelberg on the initial run.

By jumpstarting, ORSLocal takes over this initial build, so that subsequent builds run flawlessly without the need to manually create the file system. This is deemed good practice by the ORS developer team (refer to the ORS reference for details).

Conversely, if you choose not to jumpstart (dry = TRUE), ORSLocal does not interact with Docker until explicitly told to. The configuration and extract files as well as the rest of the file system do not exist yet. You can manually jumpstart by calling $up(). Initializing a dry instance can be useful if you need to make changes to the compose file before building ORS for the first time.

Manual changes

ORSLocal provides a range of convenience methods to change the setup of the ORS instance. All of these methods automatically read or write their changes to disk. It is also possible to make direct changes to the (parsed) configuration and compose files. In this case, reading and writing also need to be done manually. To write changes to disk, run $update(), e.g.:

 ors$compose$parsed$services$`ors-app`$container_name <- "new-name-123"
 ors$update()

To read changes done manually to the files on disk, run $update("self").

Super class

rors::ORSInstance -> ORSLocal

Public fields

paths

List of relevant file paths for the ORS setup. Includes the top directory, compose file, config file and extract file.

version

Version of the local ORS backend

compose

Information of the compose file (docker-compose.yml). The compose file holds various settings for the Docker setup. The most important settings are included in this field:

ports: A 2×2 matrix with Docker ports

name: Name of the Docker container

memory: List with memory information on total and free system memory as well as initial and max memory allocated to the Docker instance.

image: Version of the ORS image. "latest" refers to the latest stable version. "nightly" refers to the devel version.

parsed: Parsed compose file. When making changes to this object, make sure to run $update() to apply the changes. For details, refer to the official reference.

config

Information of the configuration file (ors-config.yml). The config file holds various options about the ORS instance. This field gives details about:

profiles: A named vector of active routing profiles

parsed: Parsed configuration file. When making changes to this obhect, make sure to run $update() to apply the changes. For details, refer to the ORS reference.

extract

Information on the extract file. Contains the name and size of the selected extract file.

Methods

Public methods

ORSLocal$new()

ORSLocal$update()

ORSLocal$purge()

ORSLocal$finalize()

ORSLocal$report()

ORSLocal$set_extract()

ORSLocal$rm_extract()

ORSLocal$set_name()

ORSLocal$set_port()

ORSLocal$set_ram()

ORSLocal$set_graphbuilding()

ORSLocal$set_image()

ORSLocal$add_profiles()

ORSLocal$rm_profiles()

ORSLocal$set_endpoints()

ORSLocal$up()

ORSLocal$down()

ORSLocal$start()

ORSLocal$stop()

ORSLocal$get_image()

ORSLocal$get_container()

ORSLocal$show_logs()

ORSLocal$is_built()

ORSLocal$is_running()

ORSLocal$clone()

Inherited methods

rors::ORSInstance$get_status()

rors::ORSInstance$is_mounted()

rors::ORSInstance$is_ready()

Method new()

Initialize the ORSInstance object.

Usage

ORSLocal$new(
  dir = ".",
  version = "7.2.0",
  overwrite = FALSE,
  dry = FALSE,
  verbose = TRUE,
  prompts = interactive()
)

Arguments

dir

[character]

Custom OpenRouteService directory. If not specified, the directory will be downloaded to the system's home directory. If a directory called "openrouteservice-{version}" is present, the download will be skipped. Ignored if server is not NULL.

version

[character]

The OpenRouteService version to use. Can either be a version number (e.g. 7.2.0) or "master". Ignored if server is not NULL.

overwrite

[logical]

Whether to overwrite the current OpenRouteService directory if it exists.

dry

[logical]

Whether to start a dry run, i.e. run an instance without jumpstarting.

verbose

[logical]

Level of verbosity. If TRUE, shows informative warnings and messages, spinners, progress bars and system notifications.

prompts

[logical]

Whether to ask for permission throughout the setup. Defaults to TRUE in interactive sessions.

Method update()

Updates ORS instance. Use this to apply changes made either in the file system or to the ors_instance object itself. This method is automatically called when using any method of ors_instance that changes the ORS setup.

Usage

ORSLocal$update(what = c("fs", "self"))

Arguments

what

[character]

Whether to change the file system with changes in R or update the instance object with changes in the file system. If what = "self", parses the relevant files in the ORS directory and updates the ors_instance object. If what = "fs", updates the compose file and config file based on the changes made to the ors_instance object.

Method purge()

Purge ORS instance, i.e. take down container, (optionally) delete image, delete ORS directory, and clean up R6 class.

This method can be useful for testing and writing reproducible examples and can easily be used together with on.exit.

Usage

ORSLocal$purge(image = FALSE)

Arguments

image

[logical]

Whether to remove the docker image or keep it for other projects. The default is FALSE to prevent accidentally breaking other projects.

Method finalize()

Finalize and purge an ORS instance. Alias for $purge() that automatically executes on garbage collection.

Usage

ORSLocal$finalize()

Method report()

Prints a situation report of the ORS instance. Invokes all relevant print methods that summarize the current state of the instance object.

Usage

ORSLocal$report()

Method set_extract()

Download and set an OpenStreetMap extract for use in ORS. Wrapper for oe_get().

Usage

ORSLocal$set_extract(
  place,
  provider = "geofabrik",
  timeout = 1800,
  file = NULL,
  do_use = TRUE,
  ...
)

Arguments

place

[various]

Place name, sf/sfc/bbox object or bounding box for which to download an extract file. For details, refer to oe_match().

provider

[character/NULL]

Extract provider to download extract from. Available providers can be found by running oe_providers(). If NULL, tries all providers.

timeout

[numeric]

Timeout for extract downloads. Defaults to 30 minutes to enable longer extract downloads. The adopted timeout is the maximum of this argument and getOption("timeout").

file

[character/NULL]

Path to a local OSM extract. Can either be a full path to any OSM file or the filename of an OSM file in the data folder of ORS. If file is specified, place and provider are ignored.

do_use

[logical]

If TRUE, enables graph building with the new extract. If FALSE, does not change the compose file at all.

...

Further arguments passed to oe_get().

Method rm_extract()

Removes extract files from the data directory.

Usage

ORSLocal$rm_extract(...)

Arguments

...

File names of extract files in the data directory. All files that exist are removed.

Method set_name()

Set a name for the ORS container.

Usage

ORSLocal$set_name(name = NULL)

Arguments

name

Name for the ORS container. If NULL, generates a random name ("ors-appXXXX").

Method set_port()

Set a port for the localhost of the ORS container.

Usage

ORSLocal$set_port(port = NULL)

Arguments

port

Port to use for the container. If NULL, assigns a random port using randomPort().

Method set_ram()

Set initial and max memory that the ORS container is allowed to use.

Usage

ORSLocal$set_ram(init = NULL, max = NULL)

Arguments

init

[numeric/NULL]

Initial memory. This can change if more memory is needed. If not specified, uses max. If both are NULL, estimates memory.

max

[numeric/NULL]

Maximum memory. The container is not allowed to use more memory than this value. If not specified, uses init. If both are NULL, estimates memory.

Method set_graphbuilding()

Graph building specifies whether routing graphs should be (re-)built. Turning graph building on enables new profiles to be built or the extract to be changed but significantly increases setup time. Turn this off if you are changing configuration options that do not alter routing graphs.

Usage

ORSLocal$set_graphbuilding(mode)

Arguments

mode

[logical]

Whether to turn graph building on or off.

Method set_image()

Set version of the ORS Docker image. This should preferably be compatible with the compose version.

Usage

ORSLocal$set_image(version = NULL)

Arguments

version

[character]

Version specification of the ORS image.

Method add_profiles()

Add routing profiles to the ORS configuration. ORS only builds routing graphs for active profiles.

Usage

ORSLocal$add_profiles(...)

Arguments

...

Objects of class ors_profile or character strings. If a character string is passed, it is interpreted as ors_profile(..., template = TRUE).

Method rm_profiles()

Remove routing profiles from the ORS configuration.

Usage

ORSLocal$rm_profiles(...)

Arguments

...

Names of routing profiles to remove. "default" removes profile defaults.

Method set_endpoints()

Change endpoint-specific configurations. Specifies options that are relevant for entire API endpoints such as isochrones.

Usage

ORSLocal$set_endpoints(...)

Arguments

...

[list/NULL]

Named arguments containing the configuration for the endpoints. Available endpoits are routing, isochrones, matrix, and snap. Refer to application.yml for a list of defaults.

Method up()

Create the ORS docker container and setup the ORS backend on a local host.

Usage

ORSLocal$up(wait = TRUE, ...)

Arguments

wait

logical

Whether to run a spinner and show a system notification when the setup is completed. If FALSE releases the console after the Docker container is created. You can then check the service status using $is_ready().

...

Additional flags passed to the docker up command.

Details

The exact command run by $up() is:

docker compose -p [name] -f [compose file] up -d --no-build [...]

The -p flag allows docker to run multiple ORS containers and keep them separate. It uses the service name defined in the compose file.

If not found, $up() builds the underlying OpenRouteService docker image specified by version during the initialization of ORSLocal.

Usually in detach mode (-d), docker returns terminal control back to the user. By default, $up() blocks the console, checks for errors and notifies the user when the service setup has finished. This behavior can be suppresed by setting wait = FALSE. The service status can then be checked using $is_ready() or ors_ready(). Container logs can be accessed by typing docker logs [name] in the terminal.

Method down()

Stop and remove the ORS docker container. Use this if you want to make changes to a running ORS setup such as changing the extract or selected profiles.

Usage

ORSLocal$down()

Method start()

Start the ORS docker container.

Usage

ORSLocal$start(wait = TRUE)

Arguments

wait

logical

Whether to run a spinner and show a system notification when the setup is completed. If FALSE releases the console after the Docker container is created. You can then check the service status using $is_ready().

Method stop()

Stop the ORS docker container.

Usage

ORSLocal$stop()

Method get_image()

Retrieve technical information on the docker image used.

Usage

ORSLocal$get_image()

Method get_container()

Retrieve technical information on the docker container used.

Usage

ORSLocal$get_container()

Method show_logs()

Show container logs as returned by docker logs. Useful for debugging docker setups.

Usage

ORSLocal$show_logs(format = TRUE)

Arguments

format

[logical]

If TRUE, includes ANSI colors and adds exdents. Otherwise, trims ANSI colors. Disabling formatting increases performance, which can be useful for larger logs.

Method is_built()

Checks if the ORS container is built. You can control this state using $up() and $down().

Usage

ORSLocal$is_built()

Method is_running()

Checks if the ORS container is running. You can control this state using $start() and $stop(). Check $is_ready() to see if the ORS setup succeeded.

Usage

ORSLocal$is_running()

Method clone()

The objects of this class are cloneable with this method.

Usage

ORSLocal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Download ORS, start docker and jumpstart a default session
ors <- ors_instance("~/ex_ors", version = "7.2.0")

# Take down the newly created instance
ors$down()

# Set a new extract file
ors$set_extract("Rutland")

# Allocate 100 MB of RAM
ors$set_ram(0.1)

# Add a routing profile
walking <- ors_profile("walking")
ors$add_profiles(walking)

# Set random port
ors$set_port()

# Change project name
ors$set_name("example-ors")

# Set up ORS
ors$up()

# Check if ORS container exists and is running
ors$is_built()
ors$is_running()

# Check if ORS is ready to use
ors$is_ready()

# Stop container, e.g. to make configuration changes
ors$stop()

# Make changes to the configuration
ors$set_endpoints(matrix = list(maximum_routes = 1e+05)) # default is 2500

# Change default profile settings
default <- ors_profile(maximum_snapping_radius = -1)
ors$add_profiles(default)

# If we make manual changes to the configuration, we need to apply the
# changes explicitly
ors$config$parsed$ors$engine$profiles$car$elevation <- FALSE
ors$update() # writes the current object state to the disk

# If the compose or config files are changed on disk, the object can be
# refreshed
ors$update("self") # reads the disk state to the R object

# Adding profiles does not work when the container is still built!
# Why? Because graphs need to be built for new profiles, so the container
# must be down.
if (FALSE) {
  bike <- ors_profile("bike-road")
  ors$add_profiles(bike) # only do this when the container is down!!
}

# Additionally, graphs are only re-built if we enable graph building.
# When changing the extract, this happens automatically, but we can also
# control graph building manually, e.g. for adding new profiles.
ors$set_graphbuilding(TRUE)

# Finally, start the container again to run the newly configured service
ors$start()
}